generic table / eslint configure
