--- .eslintrc.json
@@ -1,5 +1,0 @@
-{
-  "extends": ["next/core-web-vitals", "prettier",
-    "plugin:prettier/recommended"
-  ]
-}
