+++ .eslintrc.yaml
@@ -1,0 +1,55 @@
+root: true
+parser: "@typescript-eslint/parser"
+
+env:
+  browser: true
+  es6: true
+
+extends:
+  - airbnb-typescript
+  - prettier
+  - plugin:prettier/recommended
+  - next/core-web-vitals
+
+parserOptions:
+   ecmaVersion: latest
+   sourceType: "module"
+   project:
+      - tsconfig.json
+   ecmaFeatures:
+     jsx: true
+
+plugins:
+  - unused-imports
+
+
+rules:
+   "unused-imports/no-unused-imports": "error"
+   "@typescript-eslint/no-unused-vars": off
+   "@typescript-eslint/no-explicit-any": "warn"
+
+
+# rules:
+#   - "@typescript-eslint/no-unused-vars": "off"
+
+# {
+#   "extends": [
+#     "prettier",
+#     "plugin:prettier/recommended",
+#     "next/core-web-vitals"
+#   ],
+#   "plugins": ["unused-imports"],
+#   "rules": {
+#     "@typescript-eslint/no-unused-vars": "on",
+#     "unused-imports/no-unused-imports": "error",
+#     "unused-imports/no-unused-vars": [
+#       "warn",
+#       {
+#         "vars": "all",
+#         "varsIgnorePattern": "^_",
+#         "args": "after-used",
+#         "argsIgnorePattern": "^_"
+#       }
+#     ]
+#   }
+# }
