+++ package.json
@@ -21,9 +21,6 @@
     "autoprefixer": "10.4.15",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.0.0",
-    "eslint": "8.47.0",
-    "eslint-config-next": "13.4.19",
-    "eslint-plugin-prettier": "^5.0.0",
     "framer-motion": "^10.16.1",
     "hamburger-react": "^2.5.0",
     "next": "13.4.19",
@@ -43,7 +40,18 @@
     "zustand": "^4.4.1"
   },
   "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^6.4.1",
+    "@typescript-eslint/parser": "^6.4.1",
+    "eslint": "8.47.0",
+    "eslint-config-airbnb-typescript": "^17.1.0",
+    "eslint-config-next": "13.4.19",
     "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-import": "^2.28.1",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-prettier": "^5.0.0",
+    "eslint-plugin-react": "^7.33.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-unused-imports": "^3.0.0",
     "prettier": "^3.0.2",
     "prettier-plugin-tailwindcss": "^0.5.3"
   }
